pyproject.toml
[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]

[tool.poetry]
name = "ci-jobs-trigger"
version = "0.1.0"
description = "Triggers CI jobs (openshift-ci and Jenkins) scripts"
authors = ["Meni Yakove <myakove@gmail.com>", "Ruth Netser <ruty.n4@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "ci_jobs_trigger"}]


[tool.poetry.dependencies]
python = "^3.8"
openshift-python-utilities = "^5.0.26"
requests = "^2.31.0"
semver = "^3.0.2"
flask = "^3.0.2"
python-simple-logger = "^1.0.12"
shortuuid = "^1.0.11"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "*"

[tool.poetry-dynamic-versioning]
enable = true
pattern = "((?P<epoch>\\d+)!)?(?P<base>\\d+(\\.\\d+)*)"

[tool.poetry-dynamic-versioning.substitution]
files = ["VERSION"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
